feat: undo small refactor
